// app/enums/messages.ts
export enum Messages {
    DELETE_PROJECTS = 'projects_deletes',
    PROJECTS_DELETED = 'projects_deleted',
    PROJECT_UPDATED = 'projects_update',
    START_SCANNING = 'projects_scan',
    SCAN_STARTED = 'scan_started',
    SCANNED_FOLDER = 'scanned_folder',
    FINISHED_SCANNING = 'finished_scanning',
    CANCEL_SCAN = 'cancel_scan',
    PAUSE_SCAN = 'pause_scan',
    SCAN_IDLE = 'idle',
    START_SCAN_DRIVES = 'start_scanning_drives',
    FINISHED_SCANNING_DRIVES = 'finished_scanning_drives',
    RESUME_SCAN = 'resume_scan'
}